Fix C4244 (implicit narrowing conversion)

+++ thprac/src/thprac/thprac_alcostg.cpp
@@ -142,8 +142,8 @@ namespace Alcostg {
                         else
                             thPracParam.section = *mWarp ? th_sections_cba[*mStage][*mWarp - 1][*mSection] : 0;
                         thPracParam.time = *mTime;
-                        thPracParam.beer = *mBeer;
-                        thPracParam.beer_max = *mBeerMax;
+                        thPracParam.beer = static_cast<int16_t>(*mBeer);
+                        thPracParam.beer_max = static_cast<int16_t>(*mBeerMax);
                         thPracParam.score = *mScore;
 
                         mState = 0;
